created a score board
used JavaScript html and CSS.  I created by making headers then putting it in columns. I then added buttons to the html and connected functions to it from js.
i then added multiple function that would add 1,2,and 3 to the headers. I then added CSS to style it a little.